requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
sqlalchemy
pysqlite3-binary
langgraph
python-dotenv
